+++ test/chunking-form/samples/empty-chunks/_config.js
@@ -1,0 +1,6 @@
+module.exports = {
+	description: 'empty chunk pruning',
+	options: {
+		input: ['main1.js', 'main2.js']
+	}
+};