+++ test/chunking-form/samples/import-variable-duplicates/_expected/amd/main2.js
@@ -1,4 +1,4 @@
-define(['./first.js', './head.js'], function (first, head) { 'use strict';
+define(['./first.js'], function (first) { 'use strict';