+++ test/chunking-form/samples/import-variable-duplicates/_expected/system/main2.js
@@ -1,7 +1,7 @@
-System.register(['./first.js', './head.js'], function (exports, module) {
+System.register(['./first.js'], function (exports, module) {
 	'use strict';
 	return {
-		setters: [function () {}, function () {}],
+		setters: [function () {}],
 		execute: function () {