+++ test/chunking-form/samples/preserve-modules-empty/_expected/amd/main.js
@@ -1,0 +1,5 @@
+define(['./two.js'], function (__chunk_1) { 'use strict';
+
+	window.APP = { a: __chunk_1.default };
+
+});