+++ test/chunking-form/samples/preserve-modules-empty/one.js
@@ -1,0 +1,2 @@
+export { default as a } from './two';
+export { default as b } from './three';