+++ test/chunking-form/samples/preserve-modules-non-entry-imports/_expected/cjs/main.js
@@ -1,6 +1,5 @@
 'use strict';
 
-require('./dep1.js');
 var __chunk_1 = require('./dep2.js');