+++ test/chunking-form/samples/preserve-modules-non-entry-imports/_expected/es/main.js
@@ -1,4 +1,3 @@
-import './dep1.js';
 import foo from './dep2.js';