+++ test/chunking-form/samples/reexport-shortpaths/_expected/cjs/main2.js
@@ -1,5 +1,4 @@
 'use strict';
 
-require('./chunk-81f02b0a.js');
 require('./chunk-56950b81.js');