+++ test/chunking-form/samples/reexport-shortpaths/_expected/es/main1.js
@@ -1,4 +1,3 @@
-import './chunk-aad1568d.js';
 import { a as foo } from './chunk-f419c447.js';