+++ test/chunking-form/samples/reexport-shortpaths/_expected/es/main2.js
@@ -1,2 +1,1 @@
-import './chunk-aad1568d.js';
 import './chunk-f419c447.js';